refactor: move new entry logic to hook
